copy: reduce manifests list to included platforms
Certain registries like the one implemented by Sonatype Nexus
won't accept a multi platform manifest if some of some platforms
were not copied when using the experimental --platforms filter.
i.e. it does not like that a blob/ref is listed and does not
exist in its storage.

Even without this specific check it is more logical to not list
a platform that is not synced.

This change reduces the manifests list in a multi platform scenario
with only the selected, and found, platforms.

Signed-off-by: Jeremy Clerc <j.clerc@criteo.com>